Fix feature order
Although the code should already have been deterministic, there are still
multiple places where we iterate over the features, with no clear
relationship between feature index and name. Thus, we now create a fixed
vector of feature names at the beginning of the preparation script and use
it in the following sub-routines. Furthermore, the summary JSON stores the
feature ids as well. This is necessary because HTTP /score requests do not
contain the feature name, but only the id.