Change handling of target column
Update API version from 0.1 to 0.1.1.
Instead of requiring a binary factor in a column named "target", we allow
an arbitrarily named column containing arbitrarily named class labels.
This makes feature summary plots more understandable. The change affects
the classification pre-processing and thereby the output of the
GET /features method, so we define a new version of the API.